+++ src/plugins/renderers/shaders/lambert/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Lambert Shader Module
+ *
+ * Exports Lambert shader sources for use in the asset system.
+ */
+
+import lambertVertexSource from './lambert.vert.glsl';
+import lambertFragmentSource from './lambert.frag.glsl';
+
+export { lambertVertexSource, lambertFragmentSource };