+++ src/plugins/renderers/shaders/unlit/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Unlit Shader Module
+ *
+ * Exports Unlit shader sources for use in the asset system.
+ */
+
+import unlitVertexSource from './unlit.vert.glsl';
+import unlitFragmentSource from './unlit.frag.glsl';
+
+export { unlitVertexSource, unlitFragmentSource };